Create rha_quick_hack_x264_peter.py
This is a temporary solution to some problems we had because the output or RHA was not in H.264 format and thus did not play in the web browser. The related problems have already been made issues in this repository, so in the long run we should be able to delete this file.